Refactoring the code in the 'Work' package into the 'Done' package using the 'Extract Class' and 'Modifier Separate Query' techniques
I refactored the code in the 'Work' package and moved it to the 'Done' package. I applied the 'Extract Class' technique to move certain attributes from the main class to a newly created 'Matricula' class. Furthermore, I applied the 'Separate Query from Modifier' technique to split the 'getDados' method into smaller and more specific methods within the 'Enrollment' class. These changes were made to improve the maintainability and readability of the code. As I ask in the exercise.