Inicio de segundo exercício sobre Refatoração
Analysis of the code available at: <https://ideone.com/oX3c8g>.

I must apply Refactoring to the code, using the refactoring techniques learned so far. Note that this code is a modification of the previous example. In this code, we include information about whether the course has a mandatory final paper, and whether the course is finished and in what year. If the course has a mandatory final work, this work receives different names according to the level of the course (Graduation is a Final Paper, Master is a Dissertation, etc.)