Refactoring technique applied to Exercise 2: Version 1.0
This is the refactored version of the code available at https://ideone.com/oX3c8g. Refactoring techniques were applied to improve the readability and maintainability of the code, making it more organized and easier to understand. This version includes the techniques learned so far and will need a further revision to add further improvements.