First Application in SQL: Implementing the employee table and inserting records
This commit implements the creation of the employee table in an SQL application, with fields for enrollment, name, salary, gender, department and supervisor. In addition, records were inserted into the table using the INSERT INTO command. This commit marks the beginning of the development of the application and its basic structure.