Added sample client-server application using Berkeley Sockets in Java
This commit adds a sample basic client-server application in Java using Berkeley Sockets. The code consists of a server that waits for a client to connect, reads the message sent by the client, converts the message to uppercase, and sends it back to the client. The client connects to the server, sends a message to the server, reads the message sent by the server and displays it on the screen.

The Server.java and Client.java classes have been modified to add the necessary code to create and manage network connections between the client and the server.

This example was created to help beginners understand the basics of creating a client-server application using Berkeley Sockets in Java.

Reference: https://github.com/NicolasSSantos