Application example for implementing a Stack-type data structure
This commit adds the Stack class, which extends the Vector class, for implementing a Stack-like data structure. The Stack class has push, pop and top methods to add, remove and access the top element of the stack, respectively. In addition, the class has the size method to return the composition on the stack. The code also includes an example of using the Stack class in the StackProj file.