Decimal-hexadecimal converter implementation with stack storage
The Java code above implements a converter of decimal numbers to hexadecimal base using stack concepts. The program receives a non-zero integer via the keyboard, and validates the entry. Then, it uses a stack to store the remainders of the divisions by 16 of the entered number, and, finally, it unstacks these values to print the result of the conversion in hexadecimal.