Implement Singleton pattern in Superpadrãozinho class
This commit introduces the implementation of the Singleton pattern in the Superpadrãozinho class. The incorrect code that created multiple instances of the class has been updated to create only one instance of the class that can be accessed by calling the static getInstance method. The private constructor ensures that no other instance of the class can be created. The TesteSingletonIncorrect class has been replaced with TesteSingletonCorrect to test the Singleton implementation.