Implementation of the Abstract Factory pattern in Java
Adds the implementation of the Abstract Factory pattern in Java, including the AbstractFactory, ConcreteFactory, AbstractProduct, Product and Client classes. The Herbivore, Carnivore, and Animal classes have been added as common interfaces to the product classes. The AnimalWorld class has been implemented as the client class that uses factories to create animals.