Root Model of using the 'Abstract Factory' Pattern
This commit adds a code example that implements the Abstract Factory design pattern in Java. The classes Client, AbstractProductA, AbstractProductB, AbstractFactory, ConcreteFactory1, ConcreteFactory2, ProductA1, ProductA2, ProductB1 and ProductB2 were created, following the layout in 'Documentation'

In addition, a Main.java file was added that demonstrates how to use these classes to create abstract objects at runtime, following the logic of the Abstract Factory pattern. Finally, packages were created to organize the classes in a clearer and more intuitive way.

This commit aims to provide a practical and useful example of the Abstract Factory pattern, which can be used in Java projects to create families of related objects in a simple and flexible way.