Hash table implementation for product storage
This commit adds the implementation of a Hash table in Java to store objects of type Product. The Hash table has size set to 10 and uses the collision handling method called Separate Chaining. The hash function is calculated taking into account the first three characters of the product code, which indicate the country of origin.

The Product class was created with the code, description and price attributes. Methods were implemented to obtain and define attribute values.

The HashTable class was implemented with the necessary methods to register, remove and search for products in the table. In addition, the functionality of counting the quantity of products from the same country and showing the name of the country based on the product code was added.

The TesteHash test class was created to perform tests of Hash table functionalities. Several scenarios were tested, such as registering products, excluding and searching for products, counting products by country and obtaining the name of the country.

This commit meets the proposed activity request, meeting the defined specifications and requirements.

ATTENTION:

     The product code is validated, ensuring it is 13 digits long. The description cannot be blank and the price must be greater than zero.

     The functionality of showing the country name based on the product code has been implemented.

     The method of handling collisions Separate Chaining was chosen.

     The hash function is calculated taking into account the first three characters of the product code, indicating the country of origin.