Tank battle simulation in Python using Object Orientation
In this commit, we added a tank battle simulation using object-oriented programming in Python. We created a Tank class that represents a tank in the game and implemented methods for shooting, taking damage and exploding. We then create five tanks, stored in an array, and run the battle simulation, where tanks attack each other until only one survives.