Performed the activities of Class 02 of Fundamentals of Games
Exercise 1:

Change the moving_elements.py program to move the tank in all directions, using the directional keys on the keyboard.

Change the program so that the tank only follows mouse movement.

Exercise 2:

Rotate and move the Tank (image .jpg file provided) according to the keyboard keys pressed (directional keys).